update qc read

+++ scripts/01_qc_eda.py
@@ -68,7 +68,8 @@ def main() -> None:
         # If there were non-numeric values (resulting in NaNs), you might need to handle them
         if numeric_col.notna().all():
             # If all values are now numeric and the type changed, replace the original column
-            ad.var[col] = ad.var[col].astype("Int32")
+            ad.var[col] = ad.var[col].astype(float)
+            ad.var[col] = ad.var[col].astype(int)
             conver_col.append(col)
             # print(f"Converted column '{col}' to numeric.")
         else:
